Refactor the code to movex
