+++ libs/core-util/src/lib/core-util/domain.ts
@@ -1,9 +1,5 @@
 import {
-  ClientResourceShape,
   GenericClientResourceShapeOfType,
-  GenericResource,
-  GenericResourceOfType,
-  Resource,
   ResourceIdentifierObj,
   ResourceIdentifierStr,
   WsResponseResultPayload,
