--- libs/core-util/src/lib/resource-reducer/index.ts
@@ -1,2 +1,0 @@
-export * from './types';
-export * as ClientResourceReducer from './resource-reducer';
