--- libs/core-util/src/lib/resource-reducer/util.ts
@@ -1,10 +1,0 @@
-import hash from 'object-hash';
-import { UnknownRecord } from '../core-types';
-import { isObject, keyInObject } from '../core-util';
-import { GenericAction } from './types';
-
-export const hashObject = (val: UnknownRecord) => hash.MD5(val);
-
-export const isAction = (a: unknown): a is GenericAction => {
-  return isObject(a) && keyInObject(a, 'type') && keyInObject(a, 'payload');
-};
