+++ libs/movex-server/README.md
@@ -1,11 +1,3 @@
-# movex-server
+# Movex Server
 
-This library was generated with [Nx](https://nx.dev).
-
-## Building
-
-Run `nx build movex-server` to build the library.
-
-## Running unit tests
-
-Run `nx test movex-server` to execute the unit tests via [Jest](https://jestjs.io).
+This is the backend runtime.
