+++ libs/movex/src/index.ts
@@ -1,1 +1,3 @@
-export * from './lib/movex';
+export * from './lib/MovexMaster';
+export * from './lib/MovexResource';
+export * from './lib/types';
