--- libs/movex/src/lib/movex.spec.ts
@@ -1,7 +1,0 @@
-import { movex } from './movex';
-
-describe('movex', () => {
-  it('should work', () => {
-    expect(movex()).toEqual('movex');
-  });
-});
