--- libs/movex/src/lib/movex.ts
@@ -1,3 +1,0 @@
-export function movex(): string {
-  return 'movex';
-}
