+++ libs/movex/src/specs/MovexMaster.spec.ts
@@ -1,0 +1,8 @@
+test('it works', () => {
+
+  // This is the api more or less
+  /// socket.on('action', (action) => {
+
+  // })
+
+})
