--- wiki/matterio-cloud.md
@@ -1,34 +1,0 @@
----
-noteId: "ebc77b20b1fb11edb93bc50742d65a62"
-tags: []
-
----
-
-# Matterio Cloud
-
-The Saas built on top of Matterio offering, the following features for a paid subscription:
-- [ ] hosting
-- [ ] out of the box analytics
-- [ ] ease of deployments (ala vercel with next using github/gitlab integrations)
-
-This will pair @matterio/server with the Matterio Store (Redis). In the future it could even offer persistence, identity and other out of the box as well, but for the moment is just a wrapper around @matterio/server that pairs with the store.
-
-## Marketing[marketing]
-
-### Landing Page[landing-page]
-- [ ] Using Framer 
-
-- [ ] Bright colors on a purpleish backrdound. Brilliant pink, green, blue/cyan.
-
-- [ ] For the Landing page I'm thinking to showcase a bunch of games right in the page, like a different section. Would be pretty cool if people get to play them right there, and to have the whole pipeline, go through a different country or so show on the other screen (all iframes) and present how it all works with the actions and stuff.
-
-- [ ] Also I'm thinking to take some premade sample games in other engines such as phaser or whatnot and add matterio to them to showcase the ease
-
-### Collect Interest[collect-interest]
-  - Emails from the landing page
-  - People on reddit
-
-## Launch
-
-- [ ] Javascript Weekly
-- [ ] Product Hunt
